🚚 rename: types 및 api 폴더 구조 분리
