+++ nova-app/src/features/login/types/api.ts
@@ -1,6 +1,3 @@
-/**
- * 인증 응답 DTO
- */
 export type AuthResponse = {
   accessToken: string;
   memberId: number;
@@ -8,10 +5,6 @@ export type AuthResponse = {
   name: string;
 };
 
-/**
- * OAuth 콜백 요청 파라미터
- */
 export type OAuthCallbackParams = {
   code: string;
 };
-
